Renamed CatapultStatusResource to CatapultStatusEndpoint
